pyproject.toml
[project]
name = "pysample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.6",
    "mcp>=1.25.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
[tool.ruff] 
# 1行の最大文字数 
line-length = 88 

[tool.ruff.lint] 
# チェックするルールの種類（"I" は import 順の整理） 
select = ["E", "F", "I", "W"]
